css/image-fit/js/ua-stylesheet/div/001.html

<!doctype html>
<title>div default style</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div></div>
<script>
test(function() {
  var e = document.body.firstChild;
  assert_equals(getComputedStyle(e, '').OObjectFit, 'auto');
  assert_equals(getComputedStyle(e, '').OObjectPosition, '50% 50%');
});
</script>
